Refactor model naming and locations
