+++ src/app/shared/models/activity-indicator.model.ts
@@ -1,4 +1,5 @@
 export interface ActivityIndicator {
+  id: string;
   label: string;
   count: number;
 }
