+++ src/app/shared/models/country-supporters.model.ts
@@ -1,4 +1,5 @@
 export interface CountrySupporters {
+  id: string;
   country: string;
   count: number;
 }
